ref(chat): Move agent-run executor into chat/agent module
The agent executor is the subsystem, not the execution instance: run/slice
vocabulary stays in function and type names (executeAgentRun,
AgentRunOutcome), while the module path names what lives there — the
agent. chat/agent-run.ts becomes chat/agent/index.ts as the composition
root for the upcoming phase decomposition. Pure move; no behavior change
and no exported symbol renames.

Refs #746
